define a rule for single letter variables
Since single letter variable could also be just variables, i.e inside a
function's argument definition, to differtiate them is maybe more a
semantic problem than a syntactic one.
For this reason we define a new rule for `single_letter_var` to allow
the user to highlight them as they see fit, which might be a good
compromise.

This breaks a lot of tests (as expected).